requirements.txt

python-vlc
PyQt5
yt-dlp
youtube-search-python
PyYAML
